// src/app/features/appointment-scheduler/models/enum/appointment-text.enum.ts
export enum AppointmentTextEnum {
    appointmentLocationPlaceHolderText = 'Click here to enter city or zip code',
    appointmentFirstAvailableButtonText = 'First Available',
    appointmentSelectDifferentDateButtonText = 'Select Different Date',
    appointmentNotAvailableText = 'No appointments available',
    appointmentSupplementalQuestionHeaderText = 'SUPPLEMENTAL QUESTIONS',
    appointmentReviewPageHEaderText = 'Review appointment information (click section to edit):',
    appointmentReviewPageDataMessageText = 'Msg and data rates may apply.',
    appointmentReviewPageEmailFieldText = 'Email Address',
    appointmentReviewPageSubmitButtonText = 'Submit',
    appointmentSuccessPageThankYouMessageText = 'Thank you for choosing',
    appointmentSuccessPageBookedMessageText = 'Your appointment has been booked',
    appointmentSuccessPageConfirmedMessageText = 'Your appointment has been confirmed',
    appointmentSuccessPageReSchedulerMessageText = 'Your appointment is re-scheduled',
    appointmentSuccessPageCancelledMessageText = 'Your appointment is now cancelled',
    appointmentSuccessPageDetailHeaderText = 'Appointment Details',
    appointmentSuccessPageNotificationHeaderText = 'Notification Preferences',
    appointmentPhoneNumberFieldText = 'Phone',
    appointmentSuccessPageExitButtonText = 'Finish & Exit',
    appointmentSuccessPageCancelButtonText = 'Cancel Your Appointment',
    appointmentSuccessPageModifyButtonText = 'Modify Your Appointment',
    appointmentModifyModelHeaderText = 'Multi-Factor Authentication',
    appointmentModifyModelSendCodeTitleText = 'Click here to send a verification code through SMS/Email',
    appointmentCancelButtonText = 'Cancel',
    appointmentVerifyCodeButtonText = 'Verify Code',
    appointmentResendVerifyCodeText = 'Resend Code',
    appointmentEnterCodeFieldPlaceHolderText = 'Enter Code',
    appointmentConfirmationHeaderText = 'Confirmation Number',
    appointmentServiceNotSelectedErrorMessageText = 'Please select service and continue.',
    appointmentLocationNotSelectedErrorMessageText = 'Please select location and continue.',
    appointmentSlotNotSelectedErrorMessageText = 'Please select at least one time slot.',
    appointmentRequiredFieldErrorMessageText = 'Must fill out all required fields.',
    appointmentCancelledMessageText = 'Appointment Cancelled.',
    appointmentConfirmDeleteMessageText = 'Are you sure you want to delete this appointment?',
    appointmentHOOHeaderText = 'Hours of Operations',
    appointmentServiceOfferTitleText = 'Services Offered',
    appointmentSelectLocationButtonText = 'SELECT THIS LOCATION',
    appointmentRadiusHeaderText = 'Radius',
    appointmentOTPAuthenticationMessageText = 'Authentication code sent successfully. Please check your email or sms.',
    appointmentOTPRequiredMessageText = 'Code is required.',
    appointmentInvalidOTPMessageText = 'Invalid Code. Please try again.',
    appointmentModifyMessageText = 'Appointment Modified',
    appointmentRequiredSpecificFieldErrorMessageText = 'Field Required.',
    appointmentInvalidFieldMessageText = 'Invalid field data',
    appointmentAgreeMessageText = 'I agree to the',
    appointmentTermConditionMessageText = 'Terms & Conditions',
    appointmentMilesText = 'miles.',
    appointmentNonWorkingDaysText = 'No appointments available. Please select a different site or walk-in to a local office.',
    appointmentWrongEmailText = 'Email is invalid.',
    appointmentWrongPhoneText = 'Phone Number is invalid.',
    appointmentLocationAvailableTimeText = '24 hours',
    appointmentNotAvailableShortText = 'No Appt',
    mobileMonitorAskMoreTimeText = 'Are you sure you need more time?',
    mobileMonitorLeaveQueueText = 'are you sure you want to leave the line?',
    mobileMonitorRequeueMessageText = 'You have been requeued.',
    mobileMonitorRemoveFromQueueMessageText = 'You have been removed from queue.',
    mobileMonitorSurveyPageSubmitMessageText = 'Survey has been submitted successfully.',
    mobileMonitorSurveyNotSaveMessageText = 'Sorry! There is some error to submit survey.',

}

export enum AppointmentTextTypeEnum {
    appointmentLocationPlaceHolderType = 'appointmentLocationPlaceHolder',
    appointmentFirstAvailableButtonType = 'appointmentFirstAvailableButton',
    appointmentSelectDifferentDateButtonType = 'appointmentSelectDifferentDateButton',
    appointmentNotAvailableTextType = 'appointmentNotAvailable',
    appointmentSupplementalQuestionHeaderType = 'appointmentSupplementalQuestionHeader',
    appointmentReviewPageHeaderType = 'appointmentReviewPageHeader',
    appointmentReviewPageDataMessageType = 'appointmentReviewPageDataMessage',
    appointmentReviewPageEmailFieldType = 'appointmentReviewPageEmailField',
    appointmentReviewPageSubmitButtonType = 'appointmentReviewPageSubmitButton',
    appointmentSuccessPageThankYouMessageType = 'appointmentSuccessPageThankYouMessage',
    appointmentSuccessPageBookedMessageType = 'appointmentSuccessPageBookedMessage',
    appointmentSuccessPageConfirmedMessageType = 'appointmentSuccessPageConfirmedMessage',
    appointmentSuccessPageReSchedulerMessageType = 'appointmentSuccessPageReSchedulerMessage',
    appointmentSuccessPageCancelledMessageType = 'appointmentSuccessPageCancelledMessage',
    appointmentSuccessPageDetailHeaderType = 'appointmentSuccessPageDetailHeader',
    appointmentSuccessPageNotificationHeaderType = 'appointmentSuccessPageNotificationHeader',
    appointmentPhoneNumberFieldType = 'appointmentPhoneNumberField',
    appointmentSuccessPageExitButtonType = 'appointmentSuccessPageExitButton',
    appointmentSuccessPageCancelButtonType = 'appointmentSuccessPageCancelButton',
    appointmentSuccessPageModifyButtonType = 'appointmentSuccessPageModifyButton',
    appointmentModifyModelHeaderType = 'appointmentModifyModelHeader',
    appointmentModifyModelSendCodeTitleType = 'appointmentModifyModelSendCodeTitle',
    appointmentCancelButtonType = 'appointmentCancelButton',
    appointmentVerifyCodeButtonType = 'appointmentVerifyCodeButton',
    appointmentResendVerifyCodeType = 'appointmentResendVerifyCode',
    appointmentEnterCodeFieldPlaceHolderType = 'appointmentEnterCodeFieldPlaceHolder',
    appointmentConfirmationHeaderType = 'appointmentConfirmationHeader',
    appointmentServiceNotSelectedErrorMessageType = 'appointmentServiceNotSelectedErrorMessage',
    appointmentLocationNotSelectedErrorMessageType = 'appointmentLocationNotSelectedErrorMessage',
    appointmentSlotNotSelectedErrorMessageType = 'appointmentSlotNotSelectedErrorMessage',
    appointmentRequiredFieldErrorMessageType = 'appointmentRequiredFieldErrorMessage',
    appointmentCancelledMessageType = 'appointmentCancelledMessage',
    appointmentConfirmDeleteMessageType = 'appointmentConfirmDeleteMessage',
    appointmentHOOHeaderType = 'appointmentHOOHeader',
    appointmentServiceOfferTitleType = 'appointmentServiceOfferTitle',
    appointmentSelectLocationButtonType = 'appointmentSelectLocationButton',
    appointmentRadiusHeaderType = 'appointmentRadiusHeader',
    appointmentOTPAuthenticationMessageType = 'appointmentOTPAuthenticationMessage',
    appointmentOTPRequiredMessageType = 'appointmentOTPRequiredMessage',
    appointmentInvalidOTPMessageType = 'appointmentInvalidOTPMessage',
    appointmentModifyMessageType = 'appointmentModifyMessage',
    appointmentRequiredSpecificFieldErrorMessageType = 'appointmentRequiredSpecificFieldErrorMessage',
    appointmentInvalidFieldMessageType = 'appointmentInvalidFieldMessage',
    appointmentAgreeMessageType = 'appointmentAgreeMessage',
    appointmentTermConditionMessageType = 'appointmentTermConditionMessage',
    appointmentMilesType = 'appointmentMiles',
    appointmentNonWorkingDaysType = 'appointmentNonWorkingDays',
    appointmentWrongEmailType = 'appointmentWrongEmail',
    appointmentWrongPhoneType = 'appointmentWrongPhone',
    appointmentLocationAvailableTimeType = 'appointmentLocationAvailableTime',
    appointmentNotAvailableShortType = 'appointmentNotAvailableShort',
    mobileMonitorAskMoreTimeType = 'mobileMonitorAskMoreTime',
    mobileMonitorLeaveQueueType = 'mobileMonitorLeaveQueue',
    mobileMonitorRequeueMessageType = 'mobileMonitorRequeueMessage',
    mobileMonitorRemoveFromQueueMessageType = 'mobileMonitorRemoveFromQueueMessage',
    mobileMonitorSurveyPageSubmitMessageType = 'mobileMonitorSurveyPageSubmitMessage',
    mobileMonitorSurveyNotSaveMessageType = 'mobileMonitorSurveyNotSaveMessage',
}